+++ Editor/UI/Localized Reference/LocalizedStringPropertyDrawer.cs
@@ -293,7 +293,7 @@ namespace UnityEditor.Localization.UI
             }
 
             var provider = new StringTableSearchProvider();
-            var context = UnityEditor.Search.SearchService.CreateContext(provider);
+            var context = UnityEditor.Search.SearchService.CreateContext(provider, FilterIds.StringTableProviderFilter);
             var picker = new LocalizedReferencePicker<StringTableCollection>(context, "string table entry", data.tableReference.Property, data.tableEntryReference.Property);
             picker.Show();
         }